refactor(shared): convert 6 byte-identical test pairs to shared-import (#239)
Extract suite bodies to *.suite.ts factories under plugins/shared/__tests__/.
Both plugins' test files now import and invoke the factory — no @generated
header, no copy-sync gate. Follows detect-github-repo.suite.ts precedent.
