+++ plugins/shared/__tests__/priority-labels.suite.ts
@@ -1,0 +1,74 @@
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
+
+/**
+ * Shared test suite for PRIORITY_LABEL_MAP, PRIORITY_LABELS_SET, and syncPriorityLabel.
+ *
+ * **Caller contract — required hoisted mock.** The caller must hoist at module scope:
+ *
+ * ```ts
+ * vi.mock('../adapters/github-adapter', () => ({ updateLabels: vi.fn() }))
+ * ```
+ *
+ * Without this mock the `updateLabels` spy won't be set up before the module loads.
+ */
+export function registerPriorityLabelsSuite(opts: {
+  updateLabels: ReturnType<typeof vi.fn>
+  PRIORITY_LABEL_MAP: Record<string, string>
+  PRIORITY_LABELS_SET: Set<string>
+  syncPriorityLabel: (issueNum: number, priority: string) => Promise<void>
+}) {
+  const { updateLabels: mockUpdateLabels, PRIORITY_LABEL_MAP, PRIORITY_LABELS_SET, syncPriorityLabel } = opts
+
+  describe('PRIORITY_LABEL_MAP', () => {
+    it('maps all 4 canonical priorities to label names', () => {
+      expect(Object.keys(PRIORITY_LABEL_MAP)).toHaveLength(4)
+      expect(PRIORITY_LABEL_MAP['P0 - Urgent']).toBe('P0-critical')
+      expect(PRIORITY_LABEL_MAP['P1 - High']).toBe('P1-high')
+      expect(PRIORITY_LABEL_MAP['P2 - Medium']).toBe('P2-medium')
+      expect(PRIORITY_LABEL_MAP['P3 - Low']).toBe('P3-low')
+    })
+  })
+
+  describe('PRIORITY_LABELS_SET', () => {
+    it('contains all 4 label names derived from the map', () => {
+      expect(PRIORITY_LABELS_SET.size).toBe(4)
+      expect(PRIORITY_LABELS_SET.has('P0-critical')).toBe(true)
+      expect(PRIORITY_LABELS_SET.has('P1-high')).toBe(true)
+      expect(PRIORITY_LABELS_SET.has('P2-medium')).toBe(true)
+      expect(PRIORITY_LABELS_SET.has('P3-low')).toBe(true)
+    })
+  })
+
+  describe('syncPriorityLabel', () => {
+    beforeEach(() => {
+      vi.clearAllMocks()
+      vi.spyOn(console, 'error').mockImplementation(() => {})
+    })
+    afterEach(() => vi.restoreAllMocks())
+
+    it('calls updateLabels with correct add and remove for P1 - High', async () => {
+      mockUpdateLabels.mockResolvedValue(undefined)
+      await syncPriorityLabel(42, 'P1 - High')
+      expect(mockUpdateLabels).toHaveBeenCalledOnce()
+      const [issueNum, addLabels, removeLabels] = mockUpdateLabels.mock.calls[0]
+      expect(issueNum).toBe(42)
+      expect(addLabels).toEqual(['P1-high'])
+      expect(removeLabels).toHaveLength(3)
+      expect(removeLabels).toContain('P0-critical')
+      expect(removeLabels).toContain('P2-medium')
+      expect(removeLabels).toContain('P3-low')
+      expect(removeLabels).not.toContain('P1-high')
+    })
+
+    it('does not throw when updateLabels fails (non-fatal)', async () => {
+      mockUpdateLabels.mockRejectedValue(new Error('label error'))
+      await expect(syncPriorityLabel(42, 'P0 - Urgent')).resolves.toBeUndefined()
+      expect(console.error).toHaveBeenCalledWith(expect.stringContaining('Warning'))
+    })
+
+    it('is a no-op for invalid canonical priority', async () => {
+      await syncPriorityLabel(42, 'Invalid')
+      expect(mockUpdateLabels).not.toHaveBeenCalled()
+    })
+  })
+}
